Change function name, findAttacker -> findAttackers
