+++ packages/chess/src/core/detectCheck.js
