+++ packages/chess/src/core/index.js
@@ -10,6 +10,7 @@ export { default as computeMTByDirection } from './computeMTByDirection';
 export { default as computeCheckState } from './computeCheckState';
 export { default as computeRawMT } from './computeRawMT';
 export { default as computePossibleMT } from './computePossibleMT';
-export { default as findAttacker } from './findAttacker';
+export { default as findAttackers } from './findAttackers';
 export { default as predictPossibleCheck } from './predictPossibleCheck';
 export { default as parseNotation } from './parseNotation';
+export { default as detectCheck } from './detectCheck';
