+++ packages/chess/src/core/test/detectCheck.spec.js
